Serve and test use the same TS Loader
